requirements.txt
fastapi~=0.115.12
langchain~=0.3.23
python-dotenv~=1.1.0
starlette~=0.46.1
langchain-google-genai~=2.0.10
langchain_core
langchain_community
google-generativeai
pypdf
torch~=2.6.0
torchvision~=0.21.0
uvicorn~=0.34.0
pillow~=11.1.0
googletrans==4.0.0-rc1
python-multipart
pydantic~=2.11.3
PyPDF2~=3.0.1
protobuf~=5.29.4
langchain-community~=0.3.21
langchain-core~=0.3.51
transformers~=4.51.2
pandas~=2.2.3
numpy~=2.2.4
matplotlib~=3.10.1
seaborn~=0.13.2
joblib~=1.4.2
scikit-learn~=1.6.1
imbalanced-learn~=0.13.0